# [hashdisk] Add some real benchmarks