--- a/.travis.yml
+++ b/.travis.yml
@@ -9,6 +9,9 @@ os:
   - osx
 
 matrix:
+  allow_failures:
+    - name: "Benchmarks"
+    - name: "Benchmarks extended"
   include:
   - name: "Benchmarks"
     language: go