--- a/.travis/benchmark.sh
+++ b/.travis/benchmark.sh
@@ -13,12 +13,12 @@ go get golang.org/x/perf/cmd/benchstat
 
 # Run benchmark against current branch
 echo "Running benchmarks on PR branch $(git rev-parse HEAD)..."
-time go test -run NoTests -bench . -count 5 > /tmp/new
+time go test -timeout 0 -run NoTests -bench . -count 5 > /tmp/new
 
 # Run bebchnark against master
 echo "Running benchmarks on $TRAVIS_BRANCH branch..."
 git reset --hard $TRAVIS_BRANCH
-time go test -run NoTests -bench . -count 5 > /tmp/master
+time go test -timeout 0 -run NoTests -bench . -count 5 > /tmp/master
 
 echo "#########################################################"
 echo "Results:"