--- a/.travis/benchmark.sh
+++ b/.travis/benchmark.sh
@@ -23,4 +23,4 @@ time go test -timeout 0 -run NoTests -bench . -count 5 > /tmp/master
 echo "#########################################################"
 echo "Results:"
 benchstat /tmp/master /tmp/new | tee /tmp/diff
-go run .travis/pr_commenter.go Benchmarks /tmp/diff
+go run $(pwd)/.travis/pr_commenter.go Benchmarks /tmp/diff