--- a/.travis/benchmark.sh
+++ b/.travis/benchmark.sh
@@ -12,7 +12,8 @@ fi
 go get golang.org/x/perf/cmd/benchstat
 
 # Run benchmark against current branch
-echo "Running benchmarks on PR branch $(git rev-parse HEAD)..."
+export MERGE_COMMIT=$(git rev-parse HEAD)
+echo "Running benchmarks on PR branch ..."
 time go test -timeout 0 -run NoTests -bench . -count 5 > /tmp/new
 
 # Run bebchnark against master
@@ -20,6 +21,7 @@ echo "Running benchmarks on $TRAVIS_BRANCH branch..."
 git reset --hard $TRAVIS_BRANCH
 time go test -timeout 0 -run NoTests -bench . -count 5 > /tmp/master
 
+git reset --hard $MERGE_COMMIT
 echo "#########################################################"
 echo "Results:"
 benchstat /tmp/master /tmp/new | tee /tmp/diff