--- a/.travis/benchmark_extended.sh
+++ b/.travis/benchmark_extended.sh
@@ -22,4 +22,5 @@ time EXTENDED_BENCH_FILE=/tmp/master go test -timeout 0 -run TestExtendedBench -
 
 echo "#########################################################"
 echo "Results:"
-benchstat /tmp/master /tmp/new
+benchstat /tmp/master /tmp/new | tee /tmp/diff
+go run .travis/pr_commenter.go "Extended benchmarks" /tmp/diff