--- a/.travis/benchmark_extended.sh
+++ b/.travis/benchmark_extended.sh
@@ -13,12 +13,12 @@ go get golang.org/x/perf/cmd/benchstat
 
 # Run benchmark against current branch
 echo "Running benchmarks on PR branch $(git rev-parse HEAD)..."
-time EXTENDED_BENCH_FILE=/tmp/new go test -timeout 0 -run TestExtendedBench -v -count 5
+time EXTENDED_BENCH_FILE=/tmp/new go test -timeout 0 -run TestExtendedBench -v
 
 # Run bebchnark against master
 echo "Running benchmarks on $TRAVIS_BRANCH branch..."
 git reset --hard $TRAVIS_BRANCH
-time EXTENDED_BENCH_FILE=/tmp/master go test -timeout 0 -run TestExtendedBench -v -count 5
+time EXTENDED_BENCH_FILE=/tmp/master go test -timeout 0 -run TestExtendedBench -v
 
 echo "#########################################################"
 echo "Results:"