--- a/.travis/benchmark_extended.sh
+++ b/.travis/benchmark_extended.sh
@@ -12,7 +12,8 @@ fi
 go get golang.org/x/perf/cmd/benchstat
 
 # Run benchmark against current branch
-echo "Running benchmarks on PR branch $(git rev-parse HEAD)..."
+export MERGE_COMMIT=$(git rev-parse HEAD)
+echo "Running extended benchmarks on PR branch ..."
 time EXTENDED_BENCH_FILE=/tmp/new go test -timeout 0 -run TestExtendedBench -v
 
 # Run bebchnark against master
@@ -20,6 +21,7 @@ echo "Running benchmarks on $TRAVIS_BRANCH branch..."
 git reset --hard $TRAVIS_BRANCH
 time EXTENDED_BENCH_FILE=/tmp/master go test -timeout 0 -run TestExtendedBench -v
 
+git reset --hard $MERGE_COMMIT
 echo "#########################################################"
 echo "Results:"
 benchstat /tmp/master /tmp/new | tee /tmp/diff