# Fixes

- [x] [hashdisk] Do not add another key, if the key already exist (overrides instead)
- [x] [hashdisk] Only add to the totalItems (== increase load) when it's actually a new key

# Benchmarks
- [x] [util] Add ExtendedBench mode that can basically run benchmarks but with fixed `N`
- [x] [hashdisk] Fix existing benchmarks to not create a new file (instead loop around if `b.N` is bigger than the benchmark file size)
- [x] [hashdisk] Add Set benchmark for loads `0-0.5`; `0.7-0.9`, `0.9-0.95`, `0.95-0.99`
- [x] [hashdisk] Add Get benchmark for loads `0-0.5`; `0.7-0.9`, `0.9-0.95`, `0.95-0.99`
- [x] [ci] Add a `Benchmarks extended` job to travis to test those new benchmarks
- [x] Disable Go timeouts for benchmarks