@Viq111 2018-12-18
Closes #13 
Closes #21