@Viq111 2018-12-18
<details>
 <summary>Benchmarks results </summary>

```gf
name               old time/op    new time/op    delta
HashDiskWrite-2     16.9µs ±181%     5.0µs ±15%  -70.71%  (p=0.008 n=5+5)
HashDiskRead-2       1.09µs ± 1%    1.08µs ± 0%     ~     (p=0.175 n=5+4)
KvimdRandbo-2         335ns ± 0%     354ns ± 0%   +5.66%  (p=0.008 n=5+5)
KvimdWrite-2         5.67µs ± 1%    5.71µs ± 1%     ~     (p=0.222 n=5+5)
KvimdReadSame-2       151ns ±19%     156ns ±21%     ~     (p=0.690 n=5+5)
KvimdReadRandom-2    1.57µs ±63%   5.80µs ±238%     ~     (p=0.905 n=4+5)
ValuesDiskSet-2      34.1ns ± 0%    34.1ns ± 1%     ~     (p=0.603 n=5+5)
ValuesDiskGet-2      32.6ns ± 0%    32.5ns ± 1%     ~     (p=0.286 n=5+5)

name               old speed      new speed      delta
HashDiskWrite-2    2.95MB/s ±83%  4.90MB/s ±16%  +66.28%  (p=0.008 n=5+5)
HashDiskRead-2     22.0MB/s ± 1%  22.1MB/s ± 0%     ~     (p=0.175 n=5+4)
KvimdRandbo-2       345MB/s ± 0%   327MB/s ± 0%   -5.29%  (p=0.008 n=5+5)
KvimdWrite-2       20.5MB/s ± 1%  20.3MB/s ± 1%     ~     (p=0.222 n=5+5)
KvimdReadSame-2     712MB/s ±65%   687MB/s ±64%     ~     (p=0.690 n=5+5)
KvimdReadRandom-2  66.4MB/s ±91%  58.3MB/s ±90%     ~     (p=1.000 n=5+5)
ValuesDiskSet-2     234MB/s ± 0%   235MB/s ± 0%     ~     (p=0.548 n=5+5)
ValuesDiskGet-2     245MB/s ± 0%   246MB/s ± 1%     ~     (p=0.095 n=5+5)
```