@Viq111 2018-12-18
Just the `Benchmark extended` travis ci has failed as expected (no prior benchmarks)